requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

openai>=1.6.0

chromadb
tiktoken

pypdf
sentence-transformers

streamlit

pytest
